Fixed swiftlint error by adding extension
